Cleaned up interface changes after merge

+++ tests/auto/micromouse/mazetest.h
@@ -29,11 +29,15 @@ public:
 
 TEST_F(MazeTest, TestConstructor) {
     Maze m = Maze(16,16);
-    EXPECT_EQ(m.getSizeX(), 16);
-    ASSERT_EQ(m.getSizeY(), 16);
+    // cppcheck-suppress compareBoolExpressionWithInt
+    EXPECT_EQ(m.getSizeX(), static_cast<unsigned int>(16));
+    // cppcheck-suppress compareBoolExpressionWithInt
+    ASSERT_EQ(m.getSizeY(), static_cast<unsigned int>(16));
     m = Maze(12, 4);
-    EXPECT_EQ(m.getSizeX(), 12);
-    ASSERT_EQ(m.getSizeY(), 4);
+    // cppcheck-suppress compareBoolExpressionWithInt
+    EXPECT_EQ(m.getSizeX(), static_cast<unsigned int>(12));
+    // cppcheck-suppress compareBoolExpressionWithInt
+    ASSERT_EQ(m.getSizeY(), static_cast<unsigned int>(4));
 }
 
 /**
@@ -142,7 +146,8 @@ TEST_F(MazeTest, TestAdjacentNeighborsEdgeCell) {
     //Expect cell 0,0 to have neighbors of 0,1 and 1,0
     vector<Coordinate> neighbors = maze.getNeighboringCells(0,0);
 
-    EXPECT_EQ(neighbors.size(), 2);
+    // cppcheck-suppress compareBoolExpressionWithInt
+    EXPECT_EQ(neighbors.size(), static_cast<unsigned int>(2));
 
     EXPECT_TRUE(containsCoordinate(neighbors, 0, 1));
     ASSERT_TRUE(containsCoordinate(neighbors, 1, 0));
@@ -151,7 +156,8 @@ TEST_F(MazeTest, TestAdjacentNeighborsEdgeCell) {
 TEST_F(MazeTest, TestAdjacentNeighborsMiddleCellEmpty) {
     vector<Coordinate> neighbors = maze.getNeighboringCells(middleX, middleY);
 
-    EXPECT_EQ(neighbors.size(), 4);
+    // cppcheck-suppress compareBoolExpressionWithInt
+    EXPECT_EQ(neighbors.size(), static_cast<unsigned int>(4));
 
     ASSERT_TRUE(containsCoordinate(neighbors, middleX + 1, middleY));
     ASSERT_TRUE(containsCoordinate(neighbors, middleX - 1, middleY));
@@ -165,7 +171,8 @@ TEST_F(MazeTest, TestAdjacentNeighborsMiddleCellFull) {
     }
 
     vector<Coordinate> neighbors = maze.getNeighboringCells(middleX, middleY);
-    EXPECT_EQ(neighbors.size(), 0);
+    /// cppcheck-suppress compareBoolExpressionWithInt
+    EXPECT_EQ(neighbors.size(), static_cast<unsigned int>(0));
 }
 
 TEST_F(MazeTest, TestMouseVisited) {
